🛠 메인이미지&마커 컴포넌트 분리
