+++ src/pages/index.ts
@@ -1,1 +1,2 @@
 export { default as HomeIntro } from './homeIntro/HomeIntro';
+export { default as ProductInfo } from './homeIntro/component/ProductInfo';
